fix: combine NewHomeLoanCardsCarousel and KivaCategoryCarousel components into 1 component
