+++ src/components/Contentful/HomePage/NewHomeLoansCardCarousel.vue
@@ -1,23 +1,38 @@
 <template>
-	<div>
-		<kiva-category-carousel
-			:contentful-loan-channels="contentfulLoanChannels"
-			:loan-display-settings="loanDisplaySettings"
-			:new-home-exp="true"
-		/>
-	</div>
+	<section-with-background-classic
+		:background-content="background"
+		:theme-name="themeName"
+		:vertical-padding="verticalPadding"
+	>
+		<template #content>
+			<kv-page-container>
+				<div>
+					<kiva-loan-card-carousel
+						:selected-channel="selectedChannel"
+						:loan-ids="selectedChannelLoanIds"
+					/>
+				</div>
+			</kv-page-container>
+		</template>
+	</section-with-background-classic>
 </template>
 
 <script>
-import KivaCategoryCarousel from '@/components/Homepage/HomeExp/KivaCategoryCarousel';
+import gql from 'graphql-tag';
+import KivaLoanCardCarousel from '@/components/LoanCollections/HomeExp/KivaLoanCardCarousel';
 import contentfulStylesMixin from '@/plugins/contentful-ui-setting-styles-mixin';
+import SectionWithBackgroundClassic from '@/components/Contentful/SectionWithBackgroundClassic';
+import KvPageContainer from '~/@kiva/kv-components/vue/KvPageContainer';
 
 export default {
 	name: 'NewHomeLoansCardCarousel',
 	components: {
-		KivaCategoryCarousel,
+		KvPageContainer,
+		SectionWithBackgroundClassic,
+		KivaLoanCardCarousel
 	},
 	mixins: [contentfulStylesMixin],
+	inject: ['apollo', 'cookieStore'],
 	props: {
 		/**
 		 * Content group content from Contentful
@@ -26,6 +41,13 @@ export default {
 			type: Object,
 			default: () => {},
 		},
+	},
+	data() {
+		return {
+			loanChannelData: [],
+			selectedChannel: {},
+			showCarousel: false,
+		};
 	},
 	computed: {
 		/**
@@ -56,7 +78,66 @@ export default {
 				loanLimit: uiSetting?.dataObject?.loanLimit ?? 1,
 				showViewMoreCard: uiSetting?.dataObject?.showViewMoreCard ?? false
 			};
+		},
+		combinedLoanChannelData() {
+			return this.contentfulLoanChannels.map(channel => {
+				const matchedLoanChannel = this.loanChannelData.find(lc => lc.id === channel.id);
+				return { ...matchedLoanChannel, ...channel };
+			});
+		},
+		loanChannelIds() {
+			return this.contentfulLoanChannels.map(channelSetting => {
+				return channelSetting.id;
+			});
+		},
+		loanQueryLimit() {
+			return this.loanDisplaySettings?.loanLimit ?? 1;
+		},
+		selectedChannelLoanIds() {
+			const selectedChannel = this.combinedLoanChannelData.find(channel => {
+				return this.selectedChannel?.id === channel.id;
+			});
+			return selectedChannel?.loans?.values?.map(loan => loan.id) ?? [];
+		},
+		showViewMoreCard() {
+			return this.loanDisplaySettings?.showViewMoreCard ?? false;
 		}
-	}
+	},
+	mounted() {
+		this.fetchLoanChannel();
+	},
+	methods: {
+		fetchLoanChannel() {
+			this.apollo.query({
+				query: gql`query selectedLoanCategory($loanChannelIds: [Int]!, $loanLimit: Int) {
+					lend {
+						loanChannelsById(ids: $loanChannelIds){
+							id
+							name
+							url
+							loans(limit: $loanLimit) {
+								values {
+									id
+								}
+							}
+						}
+					}
+				}`,
+				variables: {
+					loanChannelIds: this.loanChannelIds,
+					loanLimit: this.loanQueryLimit
+				},
+			}).then(result => {
+				// Set All Active Loan Channels Data
+				const loanChannels = result?.data?.lend?.loanChannelsById ?? [];
+				this.loanChannelData = loanChannels;
+				// Activate the first channel available
+				const initialChannel = this.combinedLoanChannelData[0];
+				this.selectedChannel = initialChannel;
+				// Make the carousel visible
+				this.showCarousel = true;
+			});
+		}
+	},
 };
 </script>
