--- src/components/Homepage/HomeExp/KivaCategoryCarousel.vue
@@ -1,126 +1,0 @@
-<template>
-	<section-with-background-classic
-		:background-content="background"
-		:vertical-padding="verticalPadding"
-	>
-		<template #content>
-			<kv-page-container>
-				<div>
-					<kiva-loan-card-carousel
-						:selected-channel="selectedChannel"
-						:loan-ids="selectedChannelLoanIds"
-					/>
-				</div>
-			</kv-page-container>
-		</template>
-	</section-with-background-classic>
-</template>
-
-<script>
-import gql from 'graphql-tag';
-import contentfulStylesMixin from '@/plugins/contentful-ui-setting-styles-mixin';
-import KivaLoanCardCarousel from '@/components/LoanCollections/HomeExp/KivaLoanCardCarousel';
-import SectionWithBackgroundClassic from '@/components/Contentful/SectionWithBackgroundClassic';
-import KvPageContainer from '~/@kiva/kv-components/vue/KvPageContainer';
-
-export default {
-	name: 'KivaCategoryCarousel',
-	inject: ['apollo', 'cookieStore'],
-	components: {
-		KvPageContainer,
-		SectionWithBackgroundClassic,
-		KivaLoanCardCarousel
-	},
-	mixins: [contentfulStylesMixin],
-	props: {
-		contentfulLoanChannels: {
-			type: Array,
-			default: () => [],
-		},
-		loanDisplaySettings: {
-			type: Object,
-			default: () => {}
-		},
-		newHomeExp: {
-			type: Boolean,
-			default: false
-		}
-	},
-	data() {
-		return {
-			loanChannelData: [],
-			selectedChannel: {},
-			showCarousel: false,
-		};
-	},
-	computed: {
-		/**
-		 * Extract Background content from Contentful
-		* */
-		background() {
-			return this.content?.contents?.find(({ contentType }) => {
-				return contentType ? contentType === 'background' : false;
-			});
-		},
-		combinedLoanChannelData() {
-			return this.contentfulLoanChannels.map(channel => {
-				const matchedLoanChannel = this.loanChannelData.find(lc => lc.id === channel.id);
-				return { ...matchedLoanChannel, ...channel };
-			});
-		},
-		loanChannelIds() {
-			return this.contentfulLoanChannels.map(channelSetting => {
-				return channelSetting.id;
-			});
-		},
-		loanQueryLimit() {
-			return this.loanDisplaySettings?.loanLimit ?? 1;
-		},
-		selectedChannelLoanIds() {
-			const selectedChannel = this.combinedLoanChannelData.find(channel => {
-				return this.selectedChannel?.id === channel.id;
-			});
-			return selectedChannel?.loans?.values?.map(loan => loan.id) ?? [];
-		},
-		showViewMoreCard() {
-			return this.loanDisplaySettings?.showViewMoreCard ?? false;
-		}
-	},
-	mounted() {
-		this.fetchLoanChannel();
-	},
-	methods: {
-		fetchLoanChannel() {
-			this.apollo.query({
-				query: gql`query selectedLoanCategory($loanChannelIds: [Int]!, $loanLimit: Int) {
-					lend {
-						loanChannelsById(ids: $loanChannelIds){
-							id
-							name
-							url
-							loans(limit: $loanLimit) {
-								values {
-									id
-								}
-							}
-						}
-					}
-				}`,
-				variables: {
-					loanChannelIds: this.loanChannelIds,
-					loanLimit: this.loanQueryLimit
-				},
-			}).then(result => {
-				// Set All Active Loan Channels Data
-				const loanChannels = result?.data?.lend?.loanChannelsById ?? [];
-				this.loanChannelData = loanChannels;
-				// Activate the first channel available
-				const initialChannel = this.combinedLoanChannelData[0];
-				this.selectedChannel = initialChannel;
-				// Make the carousel visible
-				this.showCarousel = true;
-			});
-		}
-	},
-};
-</script>
